Choose downloads filename carefully
Rather than have the same pathname and leafname for a downloaded file
each time, this change updates the name to be in the following format:

downloads/truncatedname-epoch.ext

Here "downloads" represents the download directory. The "truncatedname"
part represents a short, lowercase version of the name. Finally "opoch"
represents the epoch time.